refactor: errorMessage 상수 파일 위치 변경
